requirements.txt
langchain
langchain-groq
langchain-community
duckdb
streamlit
plotly
pandas
python-dotenv
reportlab
sqlalchemy
